pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ticket-to-prompt"
version = "0.1.0"
description = "Turn Jira tickets into context-rich prompts for Cursor"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "langgraph>=0.2.0",
    "litellm>=1.0.0",
    "qdrant-client>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "sentence-transformers>=3.0.0",
    "gitpython>=3.1.0",
    "httpx>=0.27.0",
    "celery[redis]>=5.3.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "testcontainers>=4.0.0",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=2.0",
    "sphinxcontrib-openapi>=0.8",
    "myst-parser>=3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring Docker (deselect with '-m not integration')",
]

[tool.setuptools.packages.find]
include = [
    "api*",
    "indexing*",
    "retrieval*",
    "prompts*",
    "workflows*",
    "storage*",
    "integrations*",
    "git_analysis*",
    "config*",
    "scripts*",
    "workers*",
    "auth*",
]
